[Xposed] Disable method inlining
Some very easy methods are inlined into the caller's code, making it impossible
to hook them (or rather, the hook won't be called). Thus, this optimization has
to be disabled for Xposed to work properly.

Increasing the version number ensures that already compiled files in the Dalvik
cache are thrown away and recompiled with the new setting.

Hopefully fixes #4.