[Xposed] Ignore invoke-virtual-quick verification errors in <clinit>
In rare cases, the target method for invoke-virtual-quick cannot be resolved
because the "this" reference is still NULL when certain code paths are
followed. ART's method verifier doesn't seem to be able to detect that
previous checks ensure that the statement is never reached without an object
reference. It marks the method (and therefore the class) as invalid/rejected
and therefore makes the complete class unusable at runtime.

So far, this was only detected in class initializers, which aren't compiled
anyway. The interpreter later checks for this != NULL and throws an exception
in this case, so it's safe to ignore this error situation.