[Xposed] Only recompile odex files for the Dalvik cache
This is true for all those odex files in /system. However, there's a
special case in Google Play Services where an odex file is generated
on the device (for caching). This file becomes invalid when a new
ROM is flashed, but when trying to recompile it, Xposed's modifications
made it use odex file instead of the orignal APK as source.

Fixes #15.